MODINVSTOR-595: Code cleanup, fix 'Response head already sent'
* Fix 'Response head already sent' when doing response.end()
* Move AbstractInstanceRecordsAPI tests from OaiPmhViewTest and
  AbstractInstanceRecordsAPIImpl into AbstractInstanceRecordsAPITest
* Add missing @Test annotations
* Test that HttpServerResponse.write has been called 300 times
* Rename succeedingHttpServerResponse to getHttpServerResponseMock
* Bump JUnit version, old version causes security warnings
